cleaned up beans
